Revert "feat(cart): add checkout functionality feature file and scenarios"
