+++ src/test/resources/features/product.feature
@@ -3,7 +3,7 @@ Feature: Product Listing and Cart Functionality
   Background:
     Given I am on the login page
     When I enter username "standard_user" and password "secret_sauce"
-    Then I should see the "products" page
+    Then I should see the products page
 
   Rule: Cart should reflect added and removed items correctly
