+++ src/test/resources/test.properties
@@ -1,8 +1,4 @@
 baseUrl=https://www.saucedemo.com/
 implicitWait=3000
-products=https://www.saucedemo.com/inventory.html
-cart=https://www.saucedemo.com/cart.html
-checkout_your_information=https://www.saucedemo.com/checkout-step-one.html
-checkout_overview=https://www.saucedemo.com/checkout-step-two.html
-checkout_complete=https://www.saucedemo.com/checkout-complete.html
+
